requirements.txt
boto3==1.23.10
botocore==1.26.10
jmespath==1.0.0
kafka-python==2.0.2
numpy==1.22.4
pandas==1.4.2
py4j==0.10.9.3
pyspark==3.2.1
python-dateutil==2.8.2
pytz==2022.1
s3transfer==0.5.2
six==1.16.0
urllib3==1.26.9
